pyproject.toml
[project]
name = "biz-doc-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "markdown>=3.10.2",
    "python-docx>=1.2.0",
    "weasyprint>=68.1",
    "xhtml2pdf>=0.2.17",
]
